parser: fix an issue where two files in a row ends the input parsing
At one point, I had the last clause in an "else", but checkpatch had
warned me it was unnecessary after a break.  I didn't think of the
implications, this actually changed the behavior to break after two
filenames had been encountered (e.g., "a<b>c d").

This goes back to what I had, in a way that does not upset checkpatch.